move util to avoid in-context import
